requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
discord.py==2.2.2
exceptiongroup==1.1.1
filelock==3.11.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
imageio==2.27.0
Jinja2==3.1.2
lazy_loader==0.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.1.1
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
PySocks==1.7.1
python-dotenv==1.0.0
PyWavelets==1.4.1
requests==2.28.2
scikit-image==0.20.0
scipy==1.9.1
selenium==4.8.3
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.11.1
tifffile==2023.4.12
torch==2.0.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
